fix :  check author when edit & delete post

+++ src/db/query/post/post.queries.js
@@ -4,8 +4,8 @@ const SQL_POST_QUERIES = {
   CREATE_POST: "INSERT INTO posts (title, author, content) VALUES (?,?,?)",
   GET_POST:
     "SELECT p.id AS id, p.title AS title, u.name AS name, p.content AS content, p.create_dt AS time, password FROM posts AS p LEFT JOIN users AS u ON p.author = u.id WHERE p.id = ?;",
-  EDIT_POST: "UPDATE posts SET title = ?, content = ? WHERE id = ?",
-  DELETE_POST: "DELETE FROM posts WHERE id = ?",
+  EDIT_POST: "UPDATE posts SET title = ?, content = ? WHERE id = ? AND author = ?",
+  DELETE_POST: "DELETE FROM posts WHERE id = ? AND author =?",
 };
 
 export default SQL_POST_QUERIES;
